# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API Documentation",
        "description": "Laravel API Swagger documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/quiz/{quizId}/{questionId}": {
            "get": {
                "tags": [
                    "Quizzes"
                ],
                "summary": "Get single question of a quiz",
                "operationId": "b05a28f303bcaa861b1cc44c3876ab0f",
                "parameters": [
                    {
                        "name": "quizId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "questionId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Question fetched"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
        "/api/quiz/upload": {
            "get": {
                "tags": [
                    "Quiz Upload"
                ],
                "summary": "List uploaded quiz PDF files",
                "description": "Yuklangan PDF fayllarni ro‘yxatini qaytaradi.",
                "operationId": "b50d46fd9adab59d7bbc69b54b0b8106",
                "responses": {
                    "200": {
                        "description": "Uploaded files retrieved successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Quiz Upload"
                ],
                "summary": "Upload a PDF file and automatically generate a quiz",
                "description": "PDF-ni yuklab, undan savollar va variantlarni avtomatik ajratib, quiz yaratadi.",
                "operationId": "79066e27b8074cb22f485129819ebe35",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "PDF fayl (max 20MB)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "questions_count": {
                                        "description": "PDF dan nechta savolni random tanlab olish",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Quiz created successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/upload-pdf": {
            "post": {
                "tags": [
                    "Quiz Upload"
                ],
                "summary": "Upload PDF file only",
                "description": "PDF faylni serverga yuklash, lekin quiz yaratmaydi.",
                "operationId": "36a9cdf303bd38a0da908cf223896807",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "PDF fayl",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "PDF uploaded successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}